policy_scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
validators==0.22.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
playwright==1.41.2 